+++ pkg/virt-api/webhooks/validating-webhook/admitters/admitters_test.go
@@ -17,7 +17,7 @@
  *
  */
 
-package validating_webhook
+package admitters
 
 import (
 	"encoding/base64"
@@ -76,10 +76,10 @@ var _ = Describe("Validating Webhook", func() {
 			VMIInformer: vmiInformer,
 		})
 		config, configMapInformer = testutils.NewFakeClusterConfig(&k8sv1.ConfigMap{})
-		vmiCreateAdmitter = &VMICreateAdmitter{clusterConfig: config}
-		vmirsAdmitter = &VMIRSAdmitter{clusterConfig: config}
-		vmsAdmitter = &VMsAdmitter{clusterConfig: config}
-		migrationCreateAdmitter = &MigrationCreateAdmitter{clusterConfig: config}
+		vmiCreateAdmitter = &VMICreateAdmitter{ClusterConfig: config}
+		vmirsAdmitter = &VMIRSAdmitter{ClusterConfig: config}
+		vmsAdmitter = &VMsAdmitter{ClusterConfig: config}
+		migrationCreateAdmitter = &MigrationCreateAdmitter{ClusterConfig: config}
 	})
 
 	AfterEach(func() {
@@ -103,7 +103,7 @@ var _ = Describe("Validating Webhook", func() {
 				},
 			}
 
-			resp := vmiCreateAdmitter.admit(ar)
+			resp := vmiCreateAdmitter.Admit(ar)
 			Expect(resp.Allowed).To(Equal(false))
 			Expect(len(resp.Result.Details.Causes)).To(Equal(1))
 			Expect(resp.Result.Details.Causes[0].Field).To(Equal("spec.domain.devices.disks[0].name"))
@@ -130,7 +130,7 @@ var _ = Describe("Validating Webhook", func() {
 					},
 				},
 			}
-			resp := vmiCreateAdmitter.admit(ar)
+			resp := vmiCreateAdmitter.Admit(ar)
 			Expect(resp.Allowed).To(Equal(false))
 			Expect(resp.Result.Message).To(ContainSubstring("no memory requested"))
 		})
@@ -195,7 +195,7 @@ var _ = Describe("Validating Webhook", func() {
 						},
 					},
 				}
-				resp := vmiCreateAdmitter.admit(ar)
+				resp := vmiCreateAdmitter.Admit(ar)
 				Expect(resp.Allowed).To(Equal(false))
 				Expect(resp.Result.Message).To(Equal(`either spec.readinessProbe.tcpSocket or spec.readinessProbe.httpGet must be set if a spec.readinessProbe is specified, either spec.livenessProbe.tcpSocket or spec.livenessProbe.httpGet must be set if a spec.livenessProbe is specified`))
 			})
@@ -231,7 +231,7 @@ var _ = Describe("Validating Webhook", func() {
 						},
 					},
 				}
-				resp := vmiCreateAdmitter.admit(ar)
+				resp := vmiCreateAdmitter.Admit(ar)
 				Expect(resp.Allowed).To(Equal(false))
 				Expect(resp.Result.Message).To(Equal(`spec.readinessProbe must have exactly one probe type set, spec.livenessProbe must have exactly one probe type set`))
 			})
@@ -265,7 +265,7 @@ var _ = Describe("Validating Webhook", func() {
 						},
 					},
 				}
-				resp := vmiCreateAdmitter.admit(ar)
+				resp := vmiCreateAdmitter.Admit(ar)
 				Expect(resp.Allowed).To(Equal(true))
 			})
 			It("should reject properly configured readiness and liveness probes if no Pod Network is present", func() {
@@ -296,7 +296,7 @@ var _ = Describe("Validating Webhook", func() {
 						},
 					},
 				}
-				resp := vmiCreateAdmitter.admit(ar)
+				resp := vmiCreateAdmitter.Admit(ar)
 				Expect(resp.Allowed).To(Equal(false))
 				Expect(resp.Result.Message).To(Equal(`spec.livenessProbe is only allowed if the Pod Network is attached, spec.readinessProbe is only allowed if the Pod Network is attached`))
 			})
@@ -323,7 +323,7 @@ var _ = Describe("Validating Webhook", func() {
 					},
 				},
 			}
-			resp := vmiCreateAdmitter.admit(ar)
+			resp := vmiCreateAdmitter.Admit(ar)
 			Expect(resp.Allowed).To(Equal(true))
 		})
 
@@ -336,7 +336,7 @@ var _ = Describe("Validating Webhook", func() {
 					},
 				},
 			}
-			resp := vmiCreateAdmitter.admit(ar)
+			resp := vmiCreateAdmitter.Admit(ar)
 			Expect(resp.Allowed).To(BeFalse())
 			Expect(resp.Result.Message).To(Equal(`.very in body is a forbidden property, spec.extremely in body is a forbidden property, spec.domain in body is required`))
 		})
@@ -361,37 +361,37 @@ var _ = Describe("Validating Webhook", func() {
 				`{"very": "unknown", "spec": { "extremely": "unknown" }}`,
 				`.very in body is a forbidden property, spec.extremely in body is a forbidden property, spec.domain in body is required`,
 				webhooks.VirtualMachineInstanceGroupVersionResource,
-				vmiCreateAdmitter.admit,
+				vmiCreateAdmitter.Admit,
 			),
 			table.Entry("VirtualMachineInstance update",
 				`{"very": "unknown", "spec": { "extremely": "unknown" }}`,
 				`.very in body is a forbidden property, spec.extremely in body is a forbidden property, spec.domain in body is required`,
 				webhooks.VirtualMachineInstanceGroupVersionResource,
-				vmiUpdateAdmitter.admit,
+				vmiUpdateAdmitter.Admit,
 			),
 			table.Entry("VirtualMachineInstancePreset creation and update",
 				`{"very": "unknown", "spec": { "extremely": "unknown" }}`,
 				`.very in body is a forbidden property, spec.extremely in body is a forbidden property, spec.selector in body is required`,
 				webhooks.VirtualMachineInstancePresetGroupVersionResource,
-				vmiPresetAdmitter.admit,
+				vmiPresetAdmitter.Admit,
 			),
 			table.Entry("Migration creation ",
 				`{"very": "unknown", "spec": { "extremely": "unknown" }}`,
 				`.very in body is a forbidden property, spec.extremely in body is a forbidden property`,
 				webhooks.MigrationGroupVersionResource,
-				migrationCreateAdmitter.admit,
+				migrationCreateAdmitter.Admit,
 			),
 			table.Entry("Migration update",
 				`{"very": "unknown", "spec": { "extremely": "unknown" }}`,
 				`.very in body is a forbidden property, spec.extremely in body is a forbidden property`,
 				webhooks.MigrationGroupVersionResource,
-				migrationCreateAdmitter.admit,
+				migrationCreateAdmitter.Admit,
 			),
 			table.Entry("VirtualMachineInstanceReplicaSet creation and update",
 				`{"very": "unknown", "spec": { "extremely": "unknown" }}`,
 				`.very in body is a forbidden property, spec.extremely in body is a forbidden property, spec.selector in body is required, spec.template in body is required`,
 				webhooks.VirtualMachineInstanceReplicaSetGroupVersionResource,
-				vmirsAdmitter.admit,
+				vmirsAdmitter.Admit,
 			),
 		)
 		It("should reject valid VirtualMachineInstance spec on update", func() {
@@ -423,7 +423,7 @@ var _ = Describe("Validating Webhook", func() {
 				},
 			}
 
-			resp := vmiUpdateAdmitter.admit(ar)
+			resp := vmiUpdateAdmitter.Admit(ar)
 			Expect(resp.Allowed).To(Equal(false))
 			Expect(len(resp.Result.Details.Causes)).To(Equal(1))
 			Expect(resp.Result.Details.Causes[0].Message).To(Equal("update of VMI object is restricted"))
@@ -442,7 +442,7 @@ var _ = Describe("Validating Webhook", func() {
 				},
 			}
 
-			resp := vmirsAdmitter.admit(ar)
+			resp := vmirsAdmitter.Admit(ar)
 			Expect(resp.Allowed).To(Equal(false))
 			Expect(resp.Result.Details.Causes).To(HaveLen(len(causes)))
 			for i, cause := range causes {
@@ -504,7 +504,7 @@ var _ = Describe("Validating Webhook", func() {
 				},
 			}
 
-			resp := vmirsAdmitter.admit(ar)
+			resp := vmirsAdmitter.Admit(ar)
 			Expect(resp.Allowed).To(Equal(true))
 		})
 	})
@@ -534,7 +534,7 @@ var _ = Describe("Validating Webhook", func() {
 				},
 			}
 
-			resp := vmsAdmitter.admit(ar)
+			resp := vmsAdmitter.Admit(ar)
 			Expect(resp.Allowed).To(Equal(false))
 			Expect(len(resp.Result.Details.Causes)).To(Equal(1))
 			Expect(resp.Result.Details.Causes[0].Field).To(Equal("spec.template.spec.domain.devices.disks[0].name"))
@@ -570,7 +570,7 @@ var _ = Describe("Validating Webhook", func() {
 				},
 			}
 
-			resp := vmsAdmitter.admit(ar)
+			resp := vmsAdmitter.Admit(ar)
 			Expect(resp.Allowed).To(Equal(true))
 		})
 
@@ -615,7 +615,7 @@ var _ = Describe("Validating Webhook", func() {
 			}
 
 			enableFeatureGate("DataVolumes")
-			resp := vmsAdmitter.admit(ar)
+			resp := vmsAdmitter.Admit(ar)
 			Expect(resp.Allowed).To(Equal(true))
 		})
 
@@ -660,7 +660,7 @@ var _ = Describe("Validating Webhook", func() {
 			}
 
 			enableFeatureGate("DataVolumes")
-			resp := vmsAdmitter.admit(ar)
+			resp := vmsAdmitter.Admit(ar)
 			Expect(resp.Allowed).To(Equal(false))
 			Expect(len(resp.Result.Details.Causes)).To(Equal(1))
 			Expect(resp.Result.Details.Causes[0].Field).To(Equal("spec.dataVolumeTemplate[0]"))
@@ -696,7 +696,7 @@ var _ = Describe("Validating Webhook", func() {
 				},
 			}
 
-			resp := vmiPresetAdmitter.admit(ar)
+			resp := vmiPresetAdmitter.Admit(ar)
 			Expect(resp.Allowed).To(Equal(false))
 			Expect(len(resp.Result.Details.Causes)).To(Equal(1))
 			Expect(resp.Result.Details.Causes[0].Field).To(Equal("spec.domain.devices.disks[0]"))
@@ -723,7 +723,7 @@ var _ = Describe("Validating Webhook", func() {
 				},
 			}
 
-			resp := vmiPresetAdmitter.admit(ar)
+			resp := vmiPresetAdmitter.Admit(ar)
 			Expect(resp.Allowed).To(Equal(true))
 		})
 	})
@@ -751,7 +751,7 @@ var _ = Describe("Validating Webhook", func() {
 				},
 			}
 
-			resp := migrationCreateAdmitter.admit(ar)
+			resp := migrationCreateAdmitter.Admit(ar)
 			Expect(resp.Allowed).To(Equal(false))
 			Expect(len(resp.Result.Details.Causes)).To(Equal(1))
 			Expect(resp.Result.Details.Causes[0].Field).To(Equal("spec.vmiName"))
@@ -784,7 +784,7 @@ var _ = Describe("Validating Webhook", func() {
 				},
 			}
 
-			resp := migrationCreateAdmitter.admit(ar)
+			resp := migrationCreateAdmitter.Admit(ar)
 			Expect(resp.Allowed).To(Equal(true))
 		})
 
@@ -815,7 +815,7 @@ var _ = Describe("Validating Webhook", func() {
 				},
 			}
 
-			resp := migrationCreateAdmitter.admit(ar)
+			resp := migrationCreateAdmitter.Admit(ar)
 			Expect(resp.Allowed).To(Equal(false))
 		})
 
@@ -851,7 +851,7 @@ var _ = Describe("Validating Webhook", func() {
 				},
 			}
 
-			resp := migrationCreateAdmitter.admit(ar)
+			resp := migrationCreateAdmitter.Admit(ar)
 			Expect(resp.Allowed).To(Equal(false))
 		})
 
@@ -887,7 +887,7 @@ var _ = Describe("Validating Webhook", func() {
 				},
 			}
 
-			resp := migrationCreateAdmitter.admit(ar)
+			resp := migrationCreateAdmitter.Admit(ar)
 			Expect(resp.Allowed).To(Equal(true))
 		})
 
@@ -919,7 +919,7 @@ var _ = Describe("Validating Webhook", func() {
 				},
 			}
 
-			resp := migrationCreateAdmitter.admit(ar)
+			resp := migrationCreateAdmitter.Admit(ar)
 			Expect(resp.Allowed).To(Equal(false))
 		})
 
@@ -963,7 +963,7 @@ var _ = Describe("Validating Webhook", func() {
 				},
 			}
 
-			resp := migrationCreateAdmitter.admit(ar)
+			resp := migrationCreateAdmitter.Admit(ar)
 			Expect(resp.Allowed).To(Equal(false))
 			Expect(resp.Result.Message).To(ContainSubstring("DisksNotLiveMigratable"))
 		})
@@ -1005,7 +1005,7 @@ var _ = Describe("Validating Webhook", func() {
 				},
 			}
 
-			resp := migrationUpdateAdmitter.admit(ar)
+			resp := migrationUpdateAdmitter.Admit(ar)
 			Expect(resp.Allowed).To(Equal(false))
 		})
 
@@ -1043,7 +1043,7 @@ var _ = Describe("Validating Webhook", func() {
 				},
 			}
 
-			resp := migrationUpdateAdmitter.admit(ar)
+			resp := migrationUpdateAdmitter.Admit(ar)
 			Expect(resp.Allowed).To(Equal(true))
 		})
 	})
