+++ tools/vms-generator/BUILD.bazel
@@ -8,7 +8,7 @@ go_library(
     deps = [
         "//pkg/api/v1:go_default_library",
         "//pkg/testutils:go_default_library",
-        "//pkg/virt-api/webhooks/validating-webhook:go_default_library",
+        "//pkg/virt-api/webhooks/validating-webhook/admitters:go_default_library",
         "//pkg/virt-config:go_default_library",
         "//tools/util:go_default_library",
         "//tools/vms-generator/utils:go_default_library",
